Actually run the playbook

+++ workspace/bootstrap.bash
@@ -5,7 +5,14 @@ if [[ $(id -u) -ne 0 ]]; then
   exit 1
 fi
 
+if [[ ! -f workspace.yml ]]; then
+  echo >&2 "unable to find workspace playbook"
+  exit 1
+fi
+
 yum install python{,-pip}
 pip install ansible==2.8.6
+
+ansible-playbook ./workspace.yml
 
 exit 0
